restructuring to ensure DOM Updates are in correct file
